refactor: move button SVGs and header component to src/components
