+++ mcq-sample-collect/src/components/button-svgs.tsx
